+++ packages/image_picker/image_picker_platform_interface/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 2.1.0
 
-* Added pickMultiImage method
+* Add `pickMultiImage` method.
 
 ## 2.0.1